requirements.txt
streamlit
python-dotenv
youtube-transcript-api
langchain
langchain-community
langchain-core
langchain-groq
faiss-cpu
sentence-transformers